+++ src/common/Pagination/usePagination.js
@@ -5,7 +5,7 @@ import {
 import paginationParamName from "./paginationParamName";
 import { useState, useEffect } from "react";
 import { selectData } from "../../features/moviesSlice";
-import { selectPeopleList } from "../../features/peopleSlice";
+import { selectPeopleList } from "../../features/popularPeople/peopleSlice";
 import { useLocation } from "react-router-dom/cjs/react-router-dom.min";
 import { useSelector } from "react-redux";
