+++ src/store.js
@@ -2,7 +2,7 @@ import { configureStore } from "@reduxjs/toolkit";
 import createSagaMiddleware from "redux-saga";
 import rootSaga from "./rootSaga";
 import moviesReducer from "./features/moviesSlice";
-import peopleReducer from "./features/peopleSlice";
+import peopleReducer from "./features/popularPeople/peopleSlice";
 import personReducer from "./features/personDetails/personSlice";
 import movieDetailsReducer from "./features/movieDetailsSlice";
